+++ src/include/Game/Ball.h
@@ -11,6 +11,6 @@ class Ball : public GameObject
 {
 public:
     Ball(Vector2 p_pos, SDL_Texture *p_tex);
-    void update(double deltaTime) override;
+    void update(float deltaTime) override;
     void onCollision(GameObject *other) override;
 };
